+++ src/Eloquent/AbstractManyRelation.php
@@ -18,7 +18,6 @@
 namespace CloudCreativity\LaravelJsonApi\Eloquent;
 
 use CloudCreativity\LaravelJsonApi\Contracts\Adapter\HasManyAdapterInterface;
-use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use Illuminate\Database\Eloquent\Model;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 
@@ -43,13 +42,8 @@ abstract class AbstractManyRelation extends AbstractRelation implements HasManyA
         }
 
         $relation = $this->getRelation($record);
-        $adapter = $this->getStore()->adapterFor($relation->getModel());
 
-        if (!$adapter instanceof AbstractAdapter) {
-            throw new RuntimeException('Expecting inverse adapter to be an Eloquent adapter.');
-        }
-
-        return $adapter->queryRelation($relation, $parameters);
+        return $this->adapterFor($relation)->queryRelation($relation, $parameters);
     }
 
     /**
@@ -60,21 +54,6 @@ abstract class AbstractManyRelation extends AbstractRelation implements HasManyA
     public function relationship($record, EncodingParametersInterface $parameters)
     {
         return $this->query($record, $parameters);
-    }
-
-    /**
-     * Does the query need to be passed to the inverse adapter?
-     *
-     * @param $record
-     * @param EncodingParametersInterface $parameters
-     * @return bool
-     */
-    protected function requiresInverseAdapter($record, EncodingParametersInterface $parameters)
-    {
-        return !empty($parameters->getFilteringParameters()) ||
-            !empty($parameters->getSortParameters()) ||
-            !empty($parameters->getPaginationParameters()) ||
-            !empty($parameters->getIncludePaths());
     }
 
 }