+++ src/Eloquent/BelongsTo.php
@@ -38,7 +38,13 @@ class BelongsTo extends AbstractRelation
      */
     public function query($record, EncodingParametersInterface $parameters)
     {
-        return $record->{$this->key};
+        if (!$this->requiresInverseAdapter($record, $parameters)) {
+            return $record->{$this->key};
+        }
+
+        $relation = $this->getRelation($record);
+
+        return $this->adapterFor($relation)->queryToOne($relation, $parameters);
     }
 
     /**